requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-cors-headers==3.10.1                                                                                             djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
djongo==1.3.6
dnspython==2.1.0
idna==3.3
install==1.3.5
Pygments==2.10.0
PyJWT==2.3.0
pymongo==3.12.1
pytz==2021.3
requests==2.26.0
sqlparse==0.2.4 
typing-extensions==3.10.0.2
urllib3==1.26.7
zebull==2.2